[#101] Rename: move Modal folder to modals folders
